Fix XML crash
*This was a crash related to numbers handling on the xml files. If a xml is directed to a number and is instead loaded with a string, it could cause a crash.
Example, on spells.xml if you had something like this:
<vocation id="Test1"/>
instead of
<vocation id="1"/>
This would result in a crash and it wouldn't warn the user that something was wrong, making it hard to be corrected. This commit should fix the crash and make it easier to find and correct these bugs.